% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{coronal_plane}
\alias{coronal_plane}
\title{Add a reference plane to a 3D atlas}
\usage{
coronal_plane(atlas, input, R = NULL)
}
\arguments{
\item{atlas}{Dataframe with xyz coordinates, hemisphere codes, and a value for each voxel.}

\item{input}{Shiny reactive input object.}

\item{R}{Shiny reactive conductor to store the plane.}
}
\value{
Atlas with roi_boundary_{x,y,z} columns defining the plane.
}
\description{
The plane can be rotated around the RL axis ("slope") and the IS axis
("rotation"), with respect to an origin that may be shifted along the AP
axis ("intecept").
}
\details{
X axis: Left-Right
Y axis: Anterior-Posterior
Z axis: Inferior-Superior
}
\examples{
\dontrun{
  coronal_plane(atlas, input, R)
}

}

