% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{coronalPlot}
\alias{coronalPlot}
\title{Plot a responsive coronal slice}
\usage{
coronalPlot(atlas_wplane, input)
}
\arguments{
\item{atlas_wplane}{Dataframe that contains information about the
posterior extent of the ATL in 3 dimensions.}

\item{input}{Shiny input object containing information from control widgets.}
}
\value{
A 1-panel figure to render and display
}
\description{
In this figure, right=right. The coronal slice will update in response to the
control wigits. Hovering over data points will update a panel containing the
aparc.2009a+aseg code at that position.
}
\examples{
\dontrun{
  output$sagitalPlot <- shiny::renderPlot({
    sagitalPlot(reactiveA(), input)
  })
}

}

