% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launchApp}
\alias{launchApp}
\title{Launch the shiny application (GUI)}
\usage{
launchApp()
}
\value{
An interactive GUI with several plots and useful information.
}
\description{
Launch the shiny application (GUI)
}
\examples{
## Not run:
\dontrun{
ShinyATLROI::launchApp()
}
## End(**Not run**)
}

