% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plane_point}
\alias{plane_point}
\title{Compute point on plane, given two other points on plane}
\usage{
plane_point(plane, x, y, z)
}
\arguments{
\item{plane}{A list containing \code{normal} and \code{origin} of plane.
And any two of the following three:}

\item{x}{Coordinate along Right-Left axis}

\item{y}{Coordinate along Anterior-Posterior axis}

\item{z}{Coordinate along Inferior-Superior axis}
}
\value{
The value for the coordinate not provided
}
\description{
Takes the output of \code{get_plane} as the definition of a plane, and values
for two coordinates. The missing third coordinate will be computed and
returned. If a vector of points is provided for each coordinate the vectors
must be the same length.
}
\examples{
\dontrun{
plane_point(plane, x=-10:10, z=1) # will return 21 values for y
plane_point(plane, x=-10:10, y=1:21) # will return 21 values for z
}
}

