% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{read_atlas}
\alias{read_atlas}
\title{Read atlas and key text files and compose useful dataframes}
\usage{
read_atlas(atlas_file, key_file = NULL, txt_has_index = FALSE,
  txt_has_ijk = TRUE)
}
\arguments{
\item{atlas_file}{Output from 3dmaskdump; space-delimited text file with one
row per voxel. The last column of each row is the voxel value. The voxel
values should correspond to indexes found in the first column of the key
file.}

\item{key_file}{aparc.a2009s+aseg_rank.niml.lt, or some subset thereof. First
column is an index, and the second column is a label. If a key file is not
provided, the atlas cannot be labeled, and atlas regions across hemispheres
cannot be grouped.}

\item{txt_has_index}{Flag to indicate whether the first column of the
\code{atlas_file} is a 1D voxel index. Default: FALSE}

\item{txt_has_ijk}{Flag to indicate whether the first three columns of the
\code{atlas_file} is are the 3D voxel indexes \code{i,j,k}. Default: TRUE}
}
\value{
A list containing the parsed and formatted atlas and key.
}
\description{
The assumption is that the atlas text file was created with 3dmaskdump
(AFNI), which writes out both voxel indexes (\code{i,j,k}) and voxel
coordinates (\code{x,y,z}). The \code{i,j,k} values are not used to within this
package, but might be included in the text file. If txt_has_ijk=TRUE, the
first 3 columns (containing the \code{i,j,k} values) will be dropped.
}
\details{
X axis: Left-Right
Y axis: Anterior-Posterior
Z axis: Inferior-Superior
}
\examples{
\dontrun{
  read_atlas(atlas_file, key_file, txt_has_index=FALSE, txt_has_ijk=TRUE)
}

}

