% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{sagital_project}
\alias{sagital_project}
\title{Project 3D sagital "shell" onto 2D sagital plane}
\usage{
sagital_project(atlas, key)
}
\arguments{
\item{atlas}{Dataframe with xyz coordinates, hemisphere codes, and a value for each voxel.}
}
\value{
Dataframe with x-axis of each hemisphere collapsed to two surfaces.
}
\description{
Project 3D sagital "shell" onto 2D sagital plane
}
\examples{
\dontrun{
  sagital_project(atlas)
}

}

