% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{sagitalPlot}
\alias{sagitalPlot}
\title{Plot sagital projections with interactive ROI reference lines}
\usage{
sagitalPlot(d, atlas_wplane, input)
}
\arguments{
\item{d}{A dataframe with yz coordinates, hemisphere codes, surface codes,
and a value for each voxel projected onto the surfaces.}

\item{atlas_wplane}{A dataframe that contains information about the
posterior extent of the ATL in 3 dimensions.}

\item{input}{Shiny input object containing information from control widgets.}
}
\value{
A 4-panel figure to render and display
}
\description{
Generates a 4 panel (hemisphere by surface [medial,lateral]) figure, where
anatomical regions are color coded. The Shiny interface controls the slope,
rotation, and position of a plane that is used to mark the posterior extent
of the anterior temporal lobe (ATL) as a custom anatomical region of
interest. Hovering over data points will update a panel containing the
aparc.2009a+aseg code at that position.
}
\examples{
\dontrun{
  output$sagitalPlot <- shiny::renderPlot({
    sagitalPlot(sagital_projection, reactiveA(),input)
  })
}

}

