% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{select_plane}
\alias{select_plane}
\title{Return portion of atlas that falls on the plane}
\usage{
select_plane(atlas_wplane, raw_index = FALSE)
}
\arguments{
\item{atlas_wplane}{Dataframe with xyz coordinates, and \code{roi_boundary_{x,y,z}}}

\item{raw_index}{Flag to instruct function to return the original
aparc.a2009s+aseg indexes, rather than the internal indexes that group
together analogous regions in each hemisphere. Default: FALSE}
}
\value{
A subset of atlas_wplane, including voxels that fall on the plane.
}
\description{
The plane will be defined by fields \code{roi_boundary_{x,y,z}}.
}

