% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{select_roi}
\alias{select_roi}
\title{Return portion of atlas that falls on or anterior to plane}
\usage{
select_roi(atlas_wplane)
}
\arguments{
\item{atlas_wplane}{A dataframe with xyz coordinates, and \code{roi_boundary_{x,y,z}}}
}
\value{
A subset of atlas_wplane, including voxels that fall on or anterior to plane.
}
\description{
The plane will be defined by fields \code{roi_boundary_{x,y,z}}. Function
assumes coordinates are in RAI order, meaning that y values on the AP-axis
are large in the back and small in the front of the brain.
}
\details{
X axis: Left-Right
Y axis: Anterior-Posterior
Z axis: Inferior-Superior
}
\examples{
\dontrun{
  coronal_plane(atlas, input, R)
}

}

