% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{select_roi}
\alias{select_roi}
\title{Return portion of atlas that falls on or anterior to plane}
\usage{
select_roi(atlas_wplane, anterior = TRUE, raw_index = FALSE)
}
\arguments{
\item{atlas_wplane}{Dataframe with xyz coordinates, and \code{roi_boundary_{x,y,z}}}

\item{anterior}{Flag indicating whether to select region anterior or posterior to the plane. Default: anterior}

\item{raw_index}{Flag to instruct function to return the original
aparc.a2009s+aseg indexes, rather than the internal indexes that group
together analogous regions in each hemisphere. Default: FALSE}
}
\value{
A subset of atlas_wplane, including voxels that fall on or anterior to plane.
}
\description{
The plane will be defined by fields \code{roi_boundary_{x,y,z}}. Function
assumes coordinates are in RAI order, meaning that y values on the AP-axis
are large in the back and small in the front of the brain.
}
\details{
X axis: Left-Right
Y axis: Anterior-Posterior
Z axis: Inferior-Superior
}
\examples{
\dontrun{
  coronal_plane(atlas, input, R)
}

}

